Restructured __init__ to singleton, added classes
