+++ examples/find_ims_documents.py
@@ -9,12 +9,11 @@ from bs4 import BeautifulSoup
 
 
 import os
-import time
 
 
 import EDGAR
 
-loader = EDGAR.dataloader(data_dir=os.path.join('..','data'), api_keys_path=os.path.join('..','api_keys.yaml'));
+loader = EDGAR.downloader(data_dir=os.path.join('..','data'));
 
 # List of companies to process
 tikrs = open(os.path.join(loader.path, os.path.join('..','tickers.txt'))).read().strip()
