+++ examples/process_nflx.py
@@ -6,7 +6,7 @@ import sys; sys.path.append('..')
 import EDGAR
 
 
-loader = EDGAR.dataloader(data_dir=os.path.join('..', 'data'), api_keys_path = os.path.join('..', 'api_keys.yaml'))
+loader = EDGAR.downloader(data_dir=os.path.join('..', 'data'))
 tikrs = ['nflx']
 
 for tikr in tikrs:
